apps/api/pyproject.toml
[project]
name = "twin-api"
version = "0.1.0"
description = "MightyTwin FastAPI backend — PostgreSQL/PostGIS storage"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy>=2.0.35",
    "geoalchemy2>=0.15.2",
    "psycopg[binary]>=3.2.0",
    "bcrypt>=4.2.0",
    "pyjwt>=2.10.0",
    "python-multipart>=0.0.20",
    "openpyxl>=3.1.5",
    "httpx>=0.27.0",
    "pytest>=8.3.0",
    "mighty_core",
    "mighty_db",
    "mighty_models",
    "mighty_spatial",
    "mighty_api",
    "mighty_migrations",
    "mighty_licensing",
]

[tool.uv.sources]
mighty_core = { workspace = true }
mighty_db = { workspace = true }
mighty_models = { workspace = true }
mighty_spatial = { workspace = true }
mighty_api = { workspace = true }
mighty_migrations = { workspace = true }
mighty_licensing = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/twin_api"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore:.*InsecureKeyLength.*:UserWarning",
]
